20w46a

+++ minecraft/src/com/mojang/realmsclient/util/task/CloseServerTask.java
@@ -40,7 +40,7 @@ public class CloseServerTask extends LongRunningTask {
 					return;
 				}
 
-				pause(var4.delaySeconds);
+				pause((long)var4.delaySeconds);
 			} catch (Exception var5) {
 				if (this.aborted()) {
 					return;
