+++ minecraft/src/com/mojang/realmsclient/util/task/OpenServerTask.java
@@ -53,7 +53,7 @@ public class OpenServerTask extends LongRunningTask {
 					return;
 				}
 
-				pause(var4.delaySeconds);
+				pause((long)var4.delaySeconds);
 			} catch (Exception var5) {
 				if (this.aborted()) {
 					return;
