+++ minecraft/src/com/mojang/realmsclient/util/task/SwitchMinigameTask.java
@@ -39,7 +39,7 @@ public class SwitchMinigameTask extends LongRunningTask {
 					return;
 				}
 
-				pause(var4.delaySeconds);
+				pause((long)var4.delaySeconds);
 			} catch (Exception var5) {
 				if (this.aborted()) {
 					return;
